A Basic TCP connection with the ability to choose the port
The TCP connection has been established and tested. The program will give the user a chance to choose the port but will choose a random one that is in the range of 1024 and 65335 if no choice is given as a fail-safe.
This project will be marked as completed for now.